server/requirements.txt
fastapi
uvicorn
python-dotenv
cohere
sentence-transformers
faiss-cpu
python-multipart

